Separate device state into a feature
Change-type: patch
